+++ casper/protocols/blockchain/block_test.go
@@ -63,7 +63,7 @@ func Test_isValidEstimate(t *testing.T) {
 }
 
 func TestBlock_isInBlockChain(t *testing.T) {
-	p := getProtocol([]uint64{10, 11})
+	p := getProtocol([]float64{10, 11})
 	p.Execute("M-0-A SJ-1-A M-1-B SJ-0-B M-0-C SJ-1-C M-1-D SJ-0-D")
 	prev := p.Msgs["A"]
 	for _, b := range []string{"B", "C", "D"} {
