+++ casper/view_test.go
@@ -7,7 +7,7 @@ import (
 )
 
 func TestView_JustificationStoreHash(t *testing.T) {
-	p := NewProtocol([]uint64{1, 2}, NewView(), nil, 10)
+	p := NewProtocol([]float64{1, 2}, NewView(), nil, 10)
 	p.Execute("M-0-A SJ-1-A M-1-B")
 	v0 := p.ValSet.GetValByName(0)
 	v1 := p.ValSet.GetValByName(1)
@@ -24,7 +24,7 @@ func TestView_JustificationStoreHash(t *testing.T) {
 }
 
 func TestView_ReceiveJustifiedMessage(t *testing.T) {
-	p := NewProtocol([]uint64{1, 2}, NewView(), nil, 10)
+	p := NewProtocol([]float64{1, 2}, NewView(), nil, 10)
 	p.Execute("M-0-A M-0-B S-1-B M-1-C")
 	v0 := p.ValSet.GetValByName(0)
 	v1 := p.ValSet.GetValByName(1)
@@ -58,7 +58,7 @@ func TestView_ReceiveJustifiedMessage(t *testing.T) {
 }
 
 func TestView_AddJustifiedMessage(t *testing.T) {
-	p := NewProtocol([]uint64{1, 2}, NewView(), nil, 10)
+	p := NewProtocol([]float64{1, 2}, NewView(), nil, 10)
 	p.Execute("M-0-A M-0-B SJ-1-A")
 	v0 := p.ValSet.GetValByName(0)
 	v1 := p.ValSet.GetValByName(1)
@@ -89,7 +89,7 @@ func TestView_AddJustifiedMessage(t *testing.T) {
 }
 
 func TestView_MultipleMessage(t *testing.T) {
-	p := NewProtocol([]uint64{1, 2}, NewView(), nil, 10)
+	p := NewProtocol([]float64{1, 2}, NewView(), nil, 10)
 	p.Execute("M-0-A SJ-1-A M-0-B M-0-C M-0-D M-0-E M-0-F S-1-F")
 	v1 := p.ValSet.GetValByName(1)
 	pendingSet := hashset.New()
